refactor: decouple Python bindings from the C++ core
- move all register_python methods and trampoline classes into a
  dedicated bindings layer (src/bindings/, one TU per group)
- core headers under include/eqlib/ no longer depend on pybind11;
  they compile as a plain C++ library
- guard the python section of the vendored hyperjet so it only
  compiles when pybind11 is included
- drop redundant explicit GIL acquisition in the lambda elements
  (pybind11's std::function caster handles this)
- make members that are exposed via Python public (Problem type
  aliases, element setters, lambda ComputeFunction)
- mark get_or_default inline (ODR violation surfaced by multiple TUs)
- PYBIND11_OVERLOAD_PURE -> PYBIND11_OVERRIDE_PURE
- splitting into six translation units also enables parallel builds
