+++ external_libraries/hyperjet/detail/hyperjet.h
@@ -825,6 +825,7 @@ public: // operators
     }
 
 public: // python
+#if defined(PYBIND11_VERSION_MAJOR)
     template <typename TModule>
     static void register_python(TModule& m, const std::string name)
     {
@@ -940,6 +941,7 @@ public: // python
             .def("__copy__", [](const Type& self) { return self; })
             .def("__deepcopy__", [](const Type& self, py::dict& memo) { return self; }, "memodict"_a);
     }
+#endif // defined(PYBIND11_VERSION_MAJOR)
 };
 
 // abs
