+++ include/eqlib/Equation.h
@@ -102,26 +102,6 @@ public: // comparison
     {
         return (size_t)this;
     }
-
-public: // python
-    template <typename TModule>
-    static void register_python(TModule& m)
-    {
-        namespace py = pybind11;
-        using namespace pybind11::literals;
-
-        using Holder = Pointer<Type>;
-
-        py::class_<Type, Holder>(m, "Equation")
-            .def(py::init<double, double, double, std::string>(), "lower_bound"_a = -infinity, "upper_bound"_a = infinity, "multiplier"_a = 0.0, "name"_a = "")
-            .def(py::init<>())
-            .def_property("is_active", &Type::is_active, &Type::set_active)
-            .def_property("lower_bound", &Type::lower_bound, &Type::set_lower_bound)
-            .def_property("upper_bound", &Type::upper_bound, &Type::set_upper_bound)
-            .def_property("name", &Type::name, &Type::set_name)
-            .def_property("multiplier", &Type::multiplier, &Type::set_multiplier)
-            .def("__repr__", &Type::to_string);
-    }
 };
 
 } // namespace eqlib
