+++ include/eqlib/LambdaConstraint.h
@@ -11,7 +11,7 @@
 namespace eqlib {
 
 class LambdaConstraint : public Constraint {
-private: // types
+public: // types
     using Type = LambdaConstraint;
     using ComputeFunction = std::function<void(const std::vector<Pointer<Equation>>&, const std::vector<Pointer<Variable>>&, Ref<Vector>, const std::vector<Ref<Vector>>&, const std::vector<Ref<Matrix>>)>;
 
@@ -33,23 +33,7 @@ public: // constructors
 public: // methods
     void compute(Ref<Vector> rs, const std::vector<Ref<Vector>>& gs, const std::vector<Ref<Matrix>>& hs) const override
     {
-        pybind11::gil_scoped_acquire acquire;
         m_compute(m_equations, m_variables, rs, gs, hs);
-    }
-
-public: // python
-    template <typename TModule>
-    static void register_python(TModule& m)
-    {
-        namespace py = pybind11;
-        using namespace pybind11::literals;
-
-        using Base = Constraint;
-        using Holder = Pointer<Type>;
-
-        py::class_<Type, Base, Holder>(m, "LambdaConstraint")
-            // constructors
-            .def(py::init<const std::vector<Pointer<Equation>>&, const std::vector<Pointer<Variable>>&, ComputeFunction>(), "equations"_a, "variables"_a, "compute"_a);
     }
 };
