+++ include/eqlib/LambdaObjective.h
@@ -10,7 +10,7 @@
 namespace eqlib {
 
 class LambdaObjective : public Objective {
-private: // types
+public: // types
     using Type = LambdaObjective;
     using ComputeFunction = std::function<double(const std::vector<Pointer<Variable>>&, Ref<Vector>, Ref<Matrix>)>;
 
@@ -30,23 +30,7 @@ public: // constructors
 public: // methods
     double compute(Ref<Vector> g, Ref<Matrix> h) const override
     {
-        pybind11::gil_scoped_acquire acquire;
         return m_compute(m_variables, g, h);
-    }
-
-public: // python
-    template <typename TModule>
-    static void register_python(TModule& m)
-    {
-        namespace py = pybind11;
-        using namespace pybind11::literals;
-
-        using Base = Objective;
-        using Holder = Pointer<Type>;
-
-        py::class_<Type, Base, Holder>(m, "LambdaObjective")
-            // constructors
-            .def(py::init<const std::vector<Pointer<Variable>>&, ComputeFunction>(), "variables"_a, "compute"_a);
     }
 };
