+++ include/eqlib/PardisoLDLT.h
@@ -170,21 +170,6 @@ public: // methods
 
         return (error != 0);
     }
-
-public: // python
-    template <typename TModule>
-    static void register_python(TModule& m)
-    {
-        namespace py = pybind11;
-        using namespace pybind11::literals;
-
-        using Base = LinearSolver;
-        using Holder = Pointer<Type>;
-
-        py::class_<Type, Base, Holder>(m, "PardisoLDLT")
-            // constructors
-            .def(py::init<>());
-    }
 };
 
 } // namespace eqlib
