+++ include/eqlib/Timer.h
@@ -36,21 +36,6 @@ public: // methods
         const Duration duration = now() - m_start;
         return duration.count();
     }
-
-public: // python
-    template <typename TModule>
-    static void register_python(TModule& m)
-    {
-        namespace py = pybind11;
-        using namespace py::literals;
-
-        py::class_<Type>(m, "Timer")
-            .def(py::init<>())
-            .def("start", &Type::start)
-            .def_property_readonly("elapsed", &Type::elapsed)
-            // deprecated misspelled alias
-            .def_property_readonly("ellapsed", &Type::elapsed);
-    }
 };
 
 } // namespace eqlib
